chore: move template lib to a dedicated folder
