+++ packages/fx-core/src/common/template-utils/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+"use strict";
+
+export * from "./templatesUtils";
+export * from "./templatesActions";
+export * from "./templates";
