+++ packages/fx-core/src/plugins/resource/bot/languageStrategy.ts
@@ -16,7 +16,7 @@ import {
   ScaffoldActionName,
   ScaffoldContext,
   scaffoldFromTemplates,
-} from "../../../common/templatesActions";
+} from "../../../common/template-utils";
 import { TeamsBotConfig } from "./configs/teamsBotConfig";
 
 export class LanguageStrategy {
